requirements.txt
fastapi
httpx
python-dotenv
cloudinary
supabase
uvicorn